metricsviews/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
dj-database-url==0.5.0
Django==2.2.24
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
google-auth==1.33.0
gunicorn==20.1.0
idna==3.2
kubernetes===18.0.0-snapshot
oauthlib==3.1.1
Pillow==8.3.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
PyJWT==2.1.0
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
Unipath==1.1
urllib3==1.26.6
websocket-client==1.1.0
whitenoise==5.3.0
